fix: restore compilation fixes - Package::new, tcp_udp transmute, Driver import, integration tests

+++ src/network/tcp_udp.rs
@@ -140,7 +140,7 @@ impl TCPConnection for SimpleSocket {
         Ok(())
     }
     fn get_state(&self) -> TCPState {
-        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst) as u32) }
+        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst)) }
     }
 }
